pyproject.toml
[project]
name = "CineGAN"
version = "0.1.0"
description = "Convert your images to cinematic tone using GAN's"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "tensorflow>=2.20.0",
]
